re-organize directory. inOneWeekend builds and runs (on macOS, even)!
