+++ README.md
@@ -1,6 +1,7 @@
-# weeker_raytracer
+weeker_raytracer
+================
 
-Based on v2 of https://github.com/RayTracing/raytracing.github.io here are some examples
+Started with v2 of https://github.com/RayTracing/raytracing.github.io here are some examples
 
 #### Rest Of Life
 
@@ -10,27 +11,32 @@ Based on v2 of https://github.com/RayTracing/raytracing.github.io here are some 
 
 #### The Next Week
 
-
 1000x1000 pixels with 2500 ray samples around each point. took over 18 hours
 
 ![final image 2](img/TNW-ch10hSH.png)
 
-
 #### In One Weekend
 
-Image took about 12.3 minutes, without   BVH. When generating same scene with BVH partitioning, took about 3 minutes.
+Image took about 12.3 minutes, without BVH. When generating same scene with BVH partitioning, took about 3 minutes.
 
 ![final image](img/IOW-ch13f.png)
 
+Build
+-----
 
-## Build
+Using `cmake`
 
 ```shell
-cd ProjectSubDirectory
-make clean && make debug
+cmake -B build
+cmake --build build
 ```
 
-## Run
+-	specify the target with the `--target <program>` option, where the program may be `inOneWeekend`, `theNextWeek`, `theRestOfYourLife`
+
+	cmake --build build --target inOneWeekend
+
+Run
+---
 
 build, then
