+++ src/InOneWeekend/hittable.h
@@ -1,7 +1,7 @@
-#ifndef HITTABLEH
-#define HITTABLEH
+#ifndef HITTABLE_H
+#define HITTABLE_H
 
-#include "ray.h"
+#include "common/rtweekend.h"
 
 class material;
