--- src/RestOfLife/Makefile
@@ -1,59 +1,0 @@
-#CXX      := -c++
-CXX      ?= -g++
-#CXXFLAGS := -pedantic-errors -Wall -Wextra -Werror
-#CXXFLAGS := -pedantic-errors -Wall -Wextra
-CXXFLAGS := -Wall -Wextra
-
-GCC_CXXFLAGS = -DMESSAGE='"Compiled with GCC"'
-CLANG_CXXFLAGS = -DMESSAGE='"Compiled with Clang"'
-UNKNOWN_CXXFLAGS = -DMESSAGE='"Compiled with an unknown compiler"'
-
-LDFLAGS  := -L/usr/lib -lstdc++ -lm
-BUILD    := ./build
-OBJ_DIR  := $(BUILD)/objects
-APP_DIR  := $(BUILD)/apps
-TARGET   := program
-INCLUDE  := -Iinclude/ -Isrc/
-
-ifeq ($(CXX),g++)
-  CXXFLAGS += $(GCC_CXXFLAGS)
-else ifeq ($(CXX),clang)
-  CXXFLAGS += $(CLANG_CXXFLAGS)
-else
-  CXXFLAGS += $(UNKNOWN_CXXFLAGS)
-endif
-
-# header files have implementations
-# program.cpp found in src/
-
-SRC      :=                     \
-	src/program.cpp             \
-	$(wildcard src/*.h)         \
-
-OBJECTS := $(SRC:%.cpp=$(OBJ_DIR)/%.o)
-
-all: build $(APP_DIR)/$(TARGET)
-
-$(OBJ_DIR)/%.o: %.cpp
-	@mkdir -p $(@D)
-	$(CXX) $(CXXFLAGS) $(INCLUDE) -o $@ -c $<
-
-$(APP_DIR)/$(TARGET): $(OBJECTS)
-	@mkdir -p $(@D)
-	$(CXX) $(CXXFLAGS) $(INCLUDE) $(LDFLAGS) -o $(APP_DIR)/$(TARGET) $(OBJECTS)
-
-.PHONY: all build clean debug release
-
-build:
-	@mkdir -p $(APP_DIR)
-	@mkdir -p $(OBJ_DIR)
-
-debug: CXXFLAGS += -DDEBUG -g
-debug: all
-
-release: CXXFLAGS += -O2
-release: all
-
-clean:
-	-@rm -rvf $(OBJ_DIR)/*
-	-@rm -rvf $(APP_DIR)/*
