--- src/RestOfLife/assets/earthmap.jpg
@@ -1,1 +1,0 @@
-../../TheNextWeek/assets/Nasa_land_ocean_ice_8192.jpg
