+++ src/RestOfLife/moving_sphere.h
