+++ src/common/rtw_stb_image.h
@@ -1,0 +1,23 @@
+#ifndef RTWEEKEND_STB_IMAGE_H
+#define RTWEEKEND_STB_IMAGE_H
+
+
+// Disable pedantic warnings for this external library.
+#ifdef _MSC_VER
+    // Microsoft Visual C++ Compiler
+    #pragma warning (push, 0)
+#endif
+
+
+
+#define STB_IMAGE_IMPLEMENTATION
+#include "common/external/stb_image.h"
+
+
+// Restore warning levels.
+#ifdef _MSC_VER
+    // Microsoft Visual C++ Compiler
+    #pragma warning (pop)
+#endif
+
+#endif
