+++ src/common/vec3.h
@@ -1,20 +1,7 @@
-#ifndef VEC3H
-#define VEC3H
-//==================================================================================================
-// Written in 2016 by Peter Shirley <ptrshrl@gmail.com>
-//
-// To the extent possible under law, the author(s) have dedicated all copyright and related and
-// neighboring rights to this software to the public domain worldwide. This software is distributed
-// without any warranty.
-//
-// You should have received a copy (see file COPYING.txt) of the CC0 Public Domain Dedication along
-// with this software. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
-//==================================================================================================
+#ifndef VEC3_H
+#define VEC3_H
 
 #include <iostream>
-#include <math.h>
-#include <stdlib.h>
-
 
 class vec3 {
     public:
